+++ spec/rubocop/cop/style/redundant_parentheses_spec.rb
@@ -90,6 +90,9 @@ describe RuboCop::Cop::Style::RedundantParentheses do
   it_behaves_like 'redundant', '(x(1, 2))', 'x(1, 2)', 'a method call'
   it_behaves_like 'redundant', '("x".to_sym)', '"x".to_sym', 'a method call'
   it_behaves_like 'redundant', '(x[:y])', 'x[:y]', 'a method call'
+  it_behaves_like 'redundant', '("foo"[0])', '"foo"[0]', 'a method call'
+  it_behaves_like 'redundant', '(["foo"][0])', '["foo"][0]', 'a method call'
+  it_behaves_like 'redundant', '({0 => :a}[0])', '{0 => :a}[0]', 'a method call'
 
   it_behaves_like 'redundant', '(!x)', '!x', 'an unary operation'
   it_behaves_like 'redundant', '(~x)', '~x', 'an unary operation'